rename: store 내부 분류 #5
